+++ miner.h
@@ -1178,7 +1178,7 @@ extern unsigned int local_work;
 extern unsigned int total_go, total_ro;
 extern const int opt_cutofftemp;
 extern int opt_log_interval;
-extern unsigned long long global_hashrate;
+extern uint64_t global_hashrate;
 extern char current_hash[68];
 extern double current_diff;
 extern uint64_t best_diff;
